Issue #379 refactor: viewModel로 로직 분리
